Move WlClient::{new, read_event} from main.rs to wl_client.rs
